Added integration tests
In addition to the manual reproduction instructions in the README, integration tests that can be run have been added. However, there is the nuance that when switching between versions of Spring Boot, some changes to the test annotations are needed due to deprecations (in 1.4.x) and removal (in 1.5.x) of classes.