Demonstrate issue with bidirectional composite key
To make clear and reproducible the issue on which I am trying to get help, this updates the `Comment` entity to use a composite key and bidirectional relationship with `Post`. In this setup, PUT requests to update adding a `Comment` to an existing `Post` fail. In the previous setup, they worked. In either setup, the JSON Patch issue remains.

An integration test that demonstrates the issue and fails is included along with a description and instructions to manually reproduce the issue.